Moving the models back into the api to get past some merge conflicts
